More constructors for `AnyValue`. Avoid using `void*` pointers.
Casts are not needed, as different pointers are stored in their
corresponding variables and not in the `void_pointer` field. Though
getters with casts from `void_pointer` to given types may be implemented
to save memory.

Callback functions of `namespace ajokki` check that each parameter has
correct type and if not, `return` without doing anything else.

All tests pass.

	modified:   cpp/ajokki/ajokki.cpp
	modified:   cpp/ajokki/ajokki_callbacks.cpp
	modified:   cpp/ylikuutio/common/any_value.cpp
	modified:   cpp/ylikuutio/common/any_value.hpp