Fix: Eigen gets downloaded before compiling any targets.
This problem was found by creating a new build directory. In the earlier
build directory Eigen had been already downloaded and when it was
downloaded the first time, after compiling targets, it was not included
in any source file so the problem was not detected that time. Creating a
new build directory and compiling there should work now following the
simple instructions given in README.

	modified:   CMakeLists.txt