Operator overload for `void operator<<` for `Tensor3` (populate tensor).
The order of dimensions for `operator<<` for `class Tensor3` from
fastest to slowest change in indices is x, y, z.

The order of dimensions for `operator<<` for `class Matrix` from
fastest to slowest change in indices is x, y.

These orders of dimensions have been chosen so that you see the `Matrix`
or the slices of `Tensor3` in your text editor with which you write C++
code, to make it easy to populate a `Matrix` or a `Tensor3`.

All tests pass.

	modified:   code/ylikuutio/linear_algebra/tensor3.cpp
	modified:   code/ylikuutio/linear_algebra/tensor3.hpp