Bugfix: `file::binary_slurp`: `std::fstream::binary`.
This fixes binary file loading in Windows, at least in Wine.

As a result, `cat1` & `cat2` are displayed correctly when running
`ajokki.exe` with `wine`: `wine ajokki.exe`.

This was not caught in any tests as tests using Google Tests do not work
in Windows yet.

All tests pass with both GCC and Clang.

	modified:   code/ylikuutio/file/file_loader.cpp