`check_and_report_if_some_string_matches`: added `const` modifier.
This function executes very slowly when compiled with Visual Studio
2017 (probably also on older versions of Visual Studio). There is no
such problem in the Linux executable compiled with GCC or with Clang,
and also there is no such problem with the cross compiled Windows
version compiled with MinGW-w64 in Linux.

This does not yet solve the problem, though. This needs to be
investigated further.

	modified:   code/ylikuutio/string/ylikuutio_string.cpp