`class Entity`: `can_be_erased` controls whether `Entity` is eraseable.
This permits naming of entities which should not be erased, such as
`AudioMaster` (`AudioMaster` is not an `Entity` yet, but will be soon).
Naming of entities which should not be erased allows calling `info` for
them and setting and getting their variables with `set` & `get`.

By default each `Entity` is not erasable. Constructors should set
`can_be_erased` to `true` when erasability is desired.

	modified:   code/ylikuutio/ontology/entity.cpp
	modified:   code/ylikuutio/ontology/entity.hpp
	modified:   code/ylikuutio/ontology/holobiont.hpp
	modified:   code/ylikuutio/ontology/material.hpp
	modified:   code/ylikuutio/ontology/movable.hpp
	modified:   code/ylikuutio/ontology/scene.hpp
	modified:   code/ylikuutio/ontology/shader.hpp
	modified:   code/ylikuutio/ontology/species.hpp
	modified:   code/ylikuutio/ontology/symbiosis.hpp
	modified:   code/ylikuutio/ontology/text3D.hpp
	modified:   code/ylikuutio/ontology/universe.cpp
	modified:   code/ylikuutio/ontology/vector_font.hpp
	modified:   code/ylikuutio/ontology/world.hpp