`yli::ontology::Scene` contains `yli::ontology::ShaderPriorityQueue`.
`yli::ontology::ShaderPriorityQueue` is a priority queue for `Shader`s
which also has a function `remove(const std::size_t childID)`.

A priority queue is needed for `Shader`s in the future so that GPGPU
`Shader`s are rendered first so that their output textures can then be
used as input textures in other `Shader`s.

Note: the output texture of a GPGPU `Shader` may be an input texture of
a later GPGPU `Shader`. The rendering order within GPGPU `Shader`s is
according to the `childID` values, the GPGPU `Shader` with the smallest
`childID` first.