Initial support for cameras (`yli::ontology::Camera`).
`Camera`s can be activated, eg. in Hirvi & Ajokki:
* in Helsinki `Scene`:
  * `activate cat_camera` (expect to see cats)
  * `activate turbo_polizei_camera` (expect to see turbo polizei)
* in Joensuu `Scene`:
  * `activate horse_camera` (expect to see a horse)

When a `Camera` is activated, if the activated `Camera` is in the active
`Scene` of the current `World` and if the current `World` is the active
`World` of the `Universe`, then the location and orientation coordinates
of the `Camera` are copied to the location and orientation coordinates
of the `Universe`.

Coordinates are still stored in the `Universe`, and coordinates specific
to a `Scene` are not implemented yet.

Support for `Camera`s is a work in progress.