`RenderMaster` owns the OpenGL context.
`Universe` does not own the OpenGL context any more.

`RenderMaster` constructor creates a hidden SDL window and binds the
OpenGL context to that hidden SDL window.

After the execution of `Universe` constructor,
`Universe::create_window_and_setup_context` can be called to create a
window and bind the OpenGL context to that window. This is currently
done in `main`.

In the future multiple windows will be supported as well.