`ApprenticeModule` provides essential apprentice functions.
`yli::ontology::ApprenticeModule` is the counterpart of
`yli::ontology::MasterModule`. Whereas `MasterModule` provides the
master functions of a master-apprentice relationship, `ApprenticeModule`
provides the apprentice functions of a master-apprentice relationship.

`MasterModule` is no more a class template. It is just a regular class.

`ApprenticeModule` is currently in use for the following
master-apprentice relationships:
* `Brain` and `Movable`.
* `Font2D` and `Console`.
* `SymbiontSpecies` and `Biont`.