Bugfix: ensure that `yli::ontology` constructors get pointer to `Scene`.
Before some constructors like `RigidBodyModule` constructor used
`virtual` function `Entity::get_scene` to get a pointer to the `Scene`.
As this was called before `Scene` was initialized, the `Entity` base
class implementation of `Entity::get_scene` was called instead of the
`override` provided by `Scene`, thus the `RigidBodyModule` of the
default `Camera` created inside `Scene` constructor got `nullptr` as
`Scene` address from the `Entity` base class implemention instead of the
`override` provided by `Scene`. This was fixed by providing the pointer
to `Scene` to `RigidBodyModule` as a constructor parameter.