`InputParametersToAnyValueCallbackWithConsole` takes `Console&`.
`InputParametersToAnyValueCallbackWithConsole` takes `Console` as a
reference.

`ConsoleCallbackObject::execute`: check `this->console_pointer` for
`nullptr` before calling the callback and converting
`this->console_pointer`  into a reference.

This change cleans up console mode callbacks' code as no checks for
`nullptr` are needed any more for `Console`, as it's up to the caller
(in this case `ConsoleCallbackObject::execute`) to provide a valid
`Console&` argument.