Add notes about Saunajuna game idea.
Saunajuna is a strategy game where you build railways, buy trains,
organize sauna train trips for customers, and in particular
organize democracy sauna train trips to different countries,
and in those you throw democracy löyly with the leaders
of different countries of the world to make the world a better place!
"saunajuna" means a sauna train in Finnish.