Use uniform buffer objects (UBOs).
UBOs enable sharing uniforms between shaders, giving more flexibility
for render pipeline design.

Also, in Vulkan all uniforms must be in uniform blocks, so UBO support
is necessary for being able to use same shaders in the future in OpenGL
and in Vulkan as well. Actual development of Vulkan support has not been
started but using UBOs is one of the necessary steps.

UBOs are a work in progress and will probably need optimization but this
is the start.

Light sources are specific to each `Scene`, not any more to each model.

`Camera` constructor activates the `Camera`. This will probably change
in the future.

Activating the `Camera` does not make it the active `Camera` in the
`Universe`, only the activate `Camera` in the `Scene` where the `Camera`
belongs to.

Add some `std::cerr` error output for trying to use Vulkan. Vulkan will
be supported in the future.