Clean up rendering code.
`Symbiosis` rendering still goes to `Biont`s through their `Holobiont`
so that rendering can be enabled or disabled on `Holobiont` level. An
alternative would be to proceed from `SymbiontSpecies` to `Biont`s and
check whether to render or it from the `Holobiont` parent of each
`Biont`. This may change in the future according to benchmarks.