Custom memory management. WORK IN PROGRESS --- DOES NOT WORK YET!
Custom memory system is implemented as `MemorySystem` class template,
that inherits `GenericMemorySystem`.

Allocations are type-specific and owned by instances of
`MemoryAllocator` class template, that inherits
`GenericMemoryAllocator`.

The actual memory regions are owned by instances of `MemoryStorage`
class template.

`Entity` has been changed for the custom memory system as follows:

`memory_allocator` member variable will store the address of a specific
`GenericMemoryAllocator` that is responsible for the memory allocation
of this particular `Entity`', and `slotID` will store the memory slot
within that allocator. Upper 32 bits of the `slotID` will be used to
identity the specific `MemoryStorage` where this `Entity` is stored, and
lower 32 bits are used as ID within that `MemoryStorage`.

The member functions of `CustomEntityFactory` have been moved to
`EntityFactory` and `CustomEntityFactory` files have been deleted.

`ModelStruct` has no more `vector_font` field. Instead `VectorFont*` is
one of the variants of the `parent` field, as appropriate.

`MockApplication` is a mock class that inherits `Application`,
to be used for testing purposes.