Custom memory system.
Custom memory system is implemented as `MemorySystem` class template,
that inherits `GenericMemorySystem`.

Allocations are type-specific and owned by instances of
`MemoryAllocator` class template, that inherits
`GenericMemoryAllocator`. By default `MemoryAllocator` works as a pool
allocator but can be specialized for the types. New pools are created
lazily.

The actual memory regions (pools) are owned by instances of
`MemoryStorage` class template.

Custom memory system is not yet in use. It will be taken into use in the
upcoming commits.