`CallbackEngine`-`ComputeTask` is a master-apprentice relationship.
`CallbackEngine` uses `GenericMasterModule` for providing the end
condition functionality.

`ComputeTask` uses `ApprenticeModule` to access the end condition
functionality provided by `CallbackEngine`....

This is necessary for proper handling of lifetimes.

Proper handling of lifetimes is essential for the custom memory system.

Custom memory system is a work in progress and will be taken into use
soon.