Hirvi: use `Object` instead of `Terrain` as a workaround.
`Terrain` is a derivative class of `Object` but it crashes. This needs
to be investigated and fixed.

Hirvi still crashes upon deallocation:

Deleting memory allocator of type 160
ERROR: `MemoryAllocator::destroy`: `storage_i` 0 is out of bounds, size is 0
ERROR: `MemoryAllocator::destroy`: `storage_i` 0 is out of bounds, size is 0
double free or corruption (fasttop)
Aborted

This means that there is some issues still. Type 160 is `Scene`.