Bugfix: rewrite `get_first_token` and unit tests for it.
Notes:

`next_i` parameter has been added. Is is mandatory. Optional parameters
do not seem to work properly between C++ and Fortran.

`'\t'` does not seem to work in Fortran. Use `achar(9)` instead.

`'\n'` does not seem to work in Fortran. Use `achar(10)` instead.

It seems that memory regions returned by a function to be deallocated by
its caller need to be `pointer` and not `allocatable`. Otherwise a
double allocation will occur on the second call.

The string to be returned to C or C++ as `char*` needs to be a character
scalar, not an array of characters. Therefore allocate it properly:
`allocate(temp_string(token_sz + 1))`

Character scalar uses double indexing.