Test for Hirvi.
Test for Hirvi still causes segfault upon deallocation.

More tests will be added soon.

`HirviApplication`: make `entity_factory` `public` so that it can be
accessed in the test code. `Application::get_generic_entity_factory` is
not enough because `entity_factory` of `HirviApplication` is of derived
type.

In Hirvi application `#include` the `entrypoint.hpp` file that contains
`main` only when not building unit tests (there is another `main` in
`test_main.cpp` for the unit tests).