Bugfix: deallocation is done using the proper the `MemoryAllocator`.
`GenericParentModule` does not keep track of the
`GenericMemoryAllocator&` any more but instead `GenericMemoryAllocator&`
is stored in each `ConstructibleModule` as a pointer. Pointer is used
instead of reference now due to the fact that the current implementation
expects that `ConstructibleModule` has a default constructor. This might
be refactored later.

This fixes double free in Hirvi upon deallocation.
https://github.com/nrz/ylikuutio/issues/45