`GenericParentModule` constructor does not receive a memory allocator.
`GenericParentModule` constructor does not need
`GenericMemoryAllocator&` any more because it is not stored in
`GenericParentModule` any more, but in `ConstructibleModule`.
That was fixed in 153edb81a5f6c72a2df9acf6e13421644fc7456e

The change was done also in the specialized parent modules:
* `ParentOfInputModesModule`
* `ParentOfPipelinesModule`

As a result, the constructors of `ParentOfInputModesModule` and
`ParentOfPipelinesModule` also do not receive any memory allocator.