Remove support for spherical coordinates.
The main reason for this in addition to cleaning up the code is to make
the upcoming change from Y up to Z up simpler and easier to implement.

Z up is needed to follow common mathematical and physical convention.

Spherical coordinates have no actual use case at the moment.

This should potentially make the code a bit faster as spherical
coordinates do not need to be computed any more.

Should a need for spherical coordinates present itself e.g. for orbital
mechanics, a new and better implementation can be done then.