Define constructors for `BrainStruct`.
This is to enforce that necessary information about `Scene` or
`CallbackEngine` should not be forgotten.

`Scene` can be provided either as `Scene*` or by providing the global
name of the `Scene` as `std::string`.

`CallbackEngine` must be provided as `CallbackEngine*` for now.
Support for providing the global name of the `CallbackEngine` as
`std::string` will be implemented soon.