`Movable` instances can be created by providing parent's global name.
This enables loose coupling between `Movable` instances and their
`Scene` parents and is also needed for YliLisp scripting language.

Providing a pointer to the `Scene` parent remains as an alternative.

`EntityFactory` searches the parent from the `Registry` of the
`Universe` instance using the global name provided as `std::string`.

Improve tests for classes derived from `Movable`.

`Camera`:
* `get_scene`, `get_number_of_children`, `get_number_of_descendants` are
  `public`.

This belongs to:
https://github.com/nrz/ylikuutio/issues/203