`Variable::get`: check `entity_parent` for `nullptr` only if needed.
This is mainly for unit tests where `entity_parent` might be `nullptr`
so an exception is thrown when only it is necessary.

This also avoids unnecessary checks when there is no read callback in
use and thus potentially makes the code minimally faster, though the
effect of this is not measured.