`Variable` instances can be created by providing parent's global name.
This enables loose coupling between `Variable` instances and their
`Entity` parents and is also needed for YliLisp scripting language.

Providing a pointer to the `Entity` parent remains as an alternative.

`EntityFactory` searches the parent from the `Registry` of the
`Universe` instance using the global name provided as `std::string`.

`VariableStruct`:
* Add a new constructor to `VariableStruct`.
* `parent` can be either `Entity*` or `std::string`.

This belongs to:
https://github.com/nrz/ylikuutio/issues/203