`CallbackObject` instances can be created by using parent's global name.
This enables loose coupling between `CallbackObject` instances and their
`CallbackEngine` parents and is also needed for YliLisp scripting
language.

Providing a pointer to the `CallbackEngine` parent remains as an
alternative.

`EntityFactory` searches the parent from the `Registry` of the
`Universe` instance using the global name provided as `std::string`.

Add tests for `CallbackObject` and `CallbackObjectStruct`.

This belongs to:
https://github.com/nrz/ylikuutio/issues/203