`CallbackParameter` instances can be created using parent's global name.
This enables loose coupling between `CallbackParameter` instances and
their `CallbackObject` parents and is also needed for YliLisp scripting
language.

Providing a pointer to the `CallbackObject` parent remains as an
alternative.

`EntityFactory` searches the parent from the `Registry` of the
`Universe` instance using the global name provided as `std::string`.

Add tests for `CallbackParameter` and `CallbackParameterStruct`.

This belongs to:
https://github.com/nrz/ylikuutio/issues/203