`Material` instances can be created by using parent's global name.
This enables loose coupling between `Material` instances and their
`Ecosystem` and `Scene` parents and is also needed for YliLisp scripting
language.

Providing a pointer to `Ecosystem` or `Scene` parent remains as an
alternative.

`EntityFactory` searches the parent from the `Registry` of the
`Universe` instance using the global name provided as `std::string`.

Add and improve tests for `Material` and `MaterialStruct`.

This belongs to:
https://github.com/nrz/ylikuutio/issues/203