`ShapeshifterTransformation` can be created using parent's global name.
This enables loose coupling between `ShapeshifterTransformation`
instances and their `Material` parents and is also needed for YliLisp
scripting language.

Providing a pointer to `Material` parent remains as an alternative.

`EntityFactory` searches the parent from the `Registry` of the
`Universe` instance using the global name provided as `std::string`.

Add and improve tests for `ShapeshifterTransformation`.

This belongs to:
https://github.com/nrz/ylikuutio/issues/203