`Text2D` instances can be created by using parent's global name.
This enables loose coupling between `Text2D` instances and their
`Font2D` parents and is also needed for YliLisp scripting language.

Providing a pointer to `Font2D` parent remains as an alternative.

`EntityFactory` searches the parent from the `Registry` of the
`Universe` instance using the global name provided as `std::string`.

Add and improve tests for `Text2D`.

This belongs to:
https://github.com/nrz/ylikuutio/issues/203