`ConsoleCallbackObject` can be created using parent's global name.
This enables loose coupling between `ConsoleCallbackObject`
instances and their `ConsoleCallbackEngine` parents and is also needed
for YliLisp scripting language.

Providing a pointer to `ConsoleCallbackEngine`parent remains as an
alternative.

`EntityFactory` searches the parent from the `Registry` of the
`Universe` instance using the global name provided as `std::string`.

Add tests for `ConsoleCallbackObject`.

This belongs to:
https://github.com/nrz/ylikuutio/issues/203