`ConsoleCallbackParameter` can be created by using parent's global name.
This enables loose coupling between `ConsoleCallbackParameter` instances
and their `ConsoleCallbackObject` parents and is also needed for YliLisp
scripting language.

Providing a pointer to `ConsoleCallbackObject` parent remains as an
alternative.

`EntityFactory` searches the parent from the `Registry` of the
`Universe` instance using the global name provided as `std::string`.

Add tests for `ConsoleCallbackParameter` and
`ConsoleCallbackParameterStruct`.

This belongs to:
https://github.com/nrz/ylikuutio/issues/203