`ComputeTask` instances can be created by using parent's global name.
This enables loose coupling between `ComputeTask` instances and their
`Pipeline` parents and is also needed for YliLisp scripting language.

Providing a pointer to `Pipeline` parent remains as an alternative.

`EntityFactory` searches the parent from the `Registry` of the
`Universe` instance using the global name provided as `std::string`.

Add and improve tests for `ComputeTask` and `ComputeTaskStruct`.

This belongs to:
https://github.com/nrz/ylikuutio/issues/203