`LispFunction` instances can be created by using parent's global name.
This enables loose coupling between `LispFunction` instances and their
`Console` parents and is also needed for YliLisp scripting language.

Providing a pointer to `Console` parent remains as an alternative.

`EntityFactory` searches the parent from the `Registry` of the
`Universe` instance using the global name provided as `std::string`.

Add and improve tests for `LispFunctionStruct`.

This completes:
https://github.com/nrz/ylikuutio/issues/203