`copy_circular_buffer_into_begin` copies circular buffer into offset 0.
`copy_circular_buffer_into_begin` copies a circular buffer into the
underlying `std::array` so that the head is copied to array offset 0.
This is done without allocating additional memory by copying element at
a time and each element maximum of n times.

Time efficiency for this is O(n^2) but the current implementation is not
intended for performance-critical use.